% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{apply_transformation}
\alias{apply_transformation}
\title{apply_transformation}
\usage{
apply_transformation(X, R, t, s)
}
\arguments{
\item{X}{a point set as an N x D matrix}

\item{R}{D x D rotation matrix}

\item{t}{1 x D translation vector}

\item{s}{scaling factor}
}
\value{
transformed point set as a N x D matrix
}
\description{
Apply rotation and translation to a point set
}
