% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ary2ps}
\alias{ary2ps}
\title{ary2ps}
\usage{
ary2ps(ary, bkg = 0)
}
\arguments{
\item{ary}{a 4d array with last dimension indexing instances.}

\item{bkg}{Extract points for array values strictly above this (default = 0)}
}
\value{
a list of point sets.
}
\description{
Convert a 4d array to a list of 3d point sets. 
The points are formed by extracting the coordinates of array values strictly above the given cut-off (default 0).
}
