% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{costWd}
\alias{costWd}
\title{costWd}
\usage{
costWd(Tr, X, Y, CX, CY, w1 = NULL, w2 = NULL, S = NULL)
}
\arguments{
\item{Tr}{Transformation vector as translation vector + rotation (angle in 2d, quaternion in 3d))}

\item{X}{matrix of means of first GMM (i.e. reference point set)}

\item{Y}{matrix of means of second GMM (i.e. moving point set)}

\item{CX}{array of covariance matrices of first GMM such that X[i,] has covariance matrix CX[,,i]}

\item{CY}{array of covariance matrices of second GMM such that Y[i,] has covariance matrix CY[,,i]}

\item{w1}{(optional) vector of mixture weights of first GMM.}

\item{w2}{(optional) vector of mixture weights of second GMM.}

\item{S}{(optional) array of pre-computed sqrtm(sqrtm(CX[,,i]) \%*\% CY[,,j] \%*\% sqrtm(CX[,,i]))}
}
\value{
cost value
}
\description{
Objective function to minimize when using GMMs
}
