% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_point_set}
\alias{crop_point_set}
\title{crop_point_set}
\usage{
crop_point_set(point.set, size, center = NULL)
}
\arguments{
\item{point.set}{a point set as a matrix with columns x,y,z.}

\item{size}{vector of distances from the target region centre along each axis. 
Points are discarded if they are outside the ellipsoid defined by size and centred on 
the given position.}

\item{center}{(optional) coordinates of the centre of the target region. If not given, 
default to the geometric median of the point set.}
}
\value{
point set as a matrix with columns x,y,z.
}
\description{
Retain points in the set that are within the given distance from the geometric median of the set.
Using the geometric median is more robust than using the centre of mass (i.e. mean).
}
