% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{dist_to_boundary}
\alias{dist_to_boundary}
\title{dist_to_boundary}
\usage{
dist_to_boundary(points, shape)
}
\arguments{
\item{points}{a data frame with x,y,z columns}

\item{shape}{an object of class ashape3d with a single alpha value}
}
\value{
vector of distances (negative values indicate points inside the shape)
}
\description{
Given a point set and an alpha-shape, get the distance of each point
to the closest boundary point of the alpha-shape.
Points inside the shape get negative values.
}
