% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_to_line}
\alias{dist_to_line}
\title{dist_to_line}
\usage{
dist_to_line(pts, a = NULL, b = NULL)
}
\arguments{
\item{pts}{a data frame or matrix with 3 columns of coordinates}

\item{a}{vector of coordinates of a point on the line}

\item{b}{a second point on the line}
}
\value{
vector of distances
}
\description{
Compute distance between a set of points and a line defined by two points
}
