% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample}
\alias{downsample}
\title{downsample}
\usage{
downsample(point.set, n = NULL, k = NULL, weights = NULL)
}
\arguments{
\item{point.set}{a point set}

\item{n}{integer, sample size.}

\item{k}{integer, number of nearest neighbours to consider to estimate local density}

\item{weights}{a vector of probability weights}
}
\value{
a point set
}
\description{
Weighted downsampling of a point set.
If point weights are not provided, they are computed to be proportional to the local density
around each point.
}
