% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_elbow}
\alias{find_elbow}
\title{find_elbow}
\usage{
find_elbow(values)
}
\arguments{
\item{values}{vector of values in decreasing order}
}
\value{
index and value of the selected point
}
\description{
Find elbow in a 2D curve represented by a list of ordered values
}
\details{
This function finds the point with maximum distance from the line between
the first and last points.
Adapted from StackOverflow:
http://stackoverflow.com/questions/2018178/finding-the-best-trade-off-point-on-a-curve
}
