% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{Gaussian_Wd}
\alias{Gaussian_Wd}
\title{Gaussian_Wd}
\usage{
Gaussian_Wd(m1, m2, S1, S2, S = NULL)
}
\arguments{
\item{m1}{mean of first distribution}

\item{m2}{mean of second distribution}

\item{S1}{variance of first distribution}

\item{S2}{variance of second distribution}

\item{S}{(optional) matrix of pre-computed sqrtm(sqrtm(S1) \%*\% S2 \%*\% sqrtm(S1))}
}
\value{
distance value
}
\description{
Compute 2-Wasserstein distance between two Gaussian distributions
}
