% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{get_shape}
\alias{get_shape}
\title{get_shape}
\usage{
get_shape(points, alpha = NULL)
}
\arguments{
\item{points}{a data frame with columns x, y, z.}

\item{alpha}{(optional) positive number}
}
\value{
an alpha-shape object of class ashape3d
}
\description{
Get the the alpha-shape of a point set.
If not given, the function automatically determines alpha using a downsampled 
point set. 
As a consequence, alpha and therefore the computed shape can vary slightly
between runs.
}
