% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{get_surface_area}
\alias{get_surface_area}
\title{get_surface_area}
\usage{
get_surface_area(as)
}
\arguments{
\item{as}{an alpha-shape object of class ashape3d}
}
\value{
a numeric value
}
\description{
Compute the surface area of an alpha-shape by summing the surfaces of the
boundary triangles
}
