% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idx2rowcol}
\alias{idx2rowcol}
\title{idx2rowcol}
\usage{
idx2rowcol(idx, n)
}
\arguments{
\item{idx}{vector of indices}

\item{n}{size of the n x n distance matrix}
}
\value{
a matrix with two columns nr and nc
}
\description{
Convert indices into a dist object to row, column coordinates of the corresponding distance matrix
}
