% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_densities}
\alias{local_densities}
\title{local_densities}
\usage{
local_densities(X, k = NULL)
}
\arguments{
\item{X}{point set, a N x D matrix}

\item{k}{(optional) number of nearest neighbors used (defaults to all points).}
}
\value{
vector of density value for each point
}
\description{
Compute local point density at each point of a point set
}
\details{
Local density is computed as in Ning X, Li F, Tian G, Wang Y (2018) 
An efficient outlier removal method for scattered point cloud data. 
PLOS ONE 13(8):e0201280. https://doi.org/10.1371/journal.pone.0201280
}
