% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locprec2cov}
\alias{locprec2cov}
\title{locprec2cov}
\usage{
locprec2cov(point.sets, scale = FALSE)
}
\arguments{
\item{point.sets}{a list of n point sets with locprec columns (locprecz column required for 3D data)}

\item{scale}{logical, whether to scale the localization precision by the variance of the coordinates}
}
\value{
a list of 2x2xn or 3x3xn arrays.
}
\description{
Converts localization precision columns to a list of arrays of covariance matrices
}
