% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locs2ps}
\alias{locs2ps}
\title{locs2ps}
\usage{
locs2ps(
  points,
  eps,
  minPts,
  keep.locprec = TRUE,
  keep.channel = TRUE,
  cluster.2d = FALSE
)
}
\arguments{
\item{points}{a point set as a data frame of coordinates with columns x,y,z.}

\item{eps}{DBSCAN parameter, size of the epsilon neighbourhood}

\item{minPts}{DBSCAN parameter, number of minimum points in the eps region}

\item{keep.locprec}{logical (default: TRUE), whether to preserve the localization precision columns}

\item{keep.channel}{logical (default: TRUE), whether to preserve channel information column}

\item{cluster.2d}{logical (default: FALSE), whether to cluster only using x,y (and ignore z)}
}
\value{
a list of matrices with columns x,y,z and eventually locprec[z] and names set to the cluster indices.
}
\description{
Cluster localizations into point sets using DBSCAN
}
