% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{multiple_registration}
\alias{multiple_registration}
\title{multiple_registration}
\usage{
multiple_registration(PS, registration, refine.iter = 20, ...)
}
\arguments{
\item{PS}{list of point sets}

\item{registration}{pairwise registration method to use}

\item{refine.iter}{Maximum number of refinement iterations (default: 20)}

\item{...}{additional arguments to the registration method}
}
\value{
a list of 
\itemize{
    \item Y: list of transformed point sets as N x d matrices
}
}
\description{
Registration of multiple point sets using tree-guided progressive registration 
followed by iterative refinement.
}
