% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_from_roi}
\alias{points_from_roi}
\title{points_from_roi}
\usage{
points_from_roi(points, roi)
}
\arguments{
\item{points}{a point set as a data frame of coordinates with columns x,y,z.}

\item{roi}{a data frame with columns x,y,z and rows min and max defining a bounding box}
}
\value{
a data frame with same columns as input
}
\description{
Extract points within given bounding box.
Points are translated so that (0,0,0) correspond to the bounding box corner defined by
roi['min',c('x','y','z')]
}
