% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps2ary}
\alias{ps2ary}
\title{ps2ary}
\usage{
ps2ary(point.sets, dims)
}
\arguments{
\item{point.sets}{a list of point sets.}

\item{dims}{vector of dimensions of the axes (x,y in 2d, x,y,z in 3d).}
}
\value{
a 3d or 4d array.
}
\description{
Convert a list of 3d point sets to a 4d array.
Also works for 2d point sets to 3d array conversion.
}
