% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{q2dr}
\alias{q2dr}
\title{Get derivative of 3D rotation matrix from quaternion}
\usage{
q2dr(q)
}
\arguments{
\item{q}{quaternion}
}
\value{
derivative of rotation matrix
}
\description{
Get derivative of 3D rotation matrix from quaternion
}
