% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{q2r}
\alias{q2r}
\title{Convert quaternion to rotation matrix
http://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation}
\usage{
q2r(q)
}
\arguments{
\item{q}{quaternion}
}
\value{
rotation matrix
}
\description{
Convert quaternion to rotation matrix
http://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation
}
