% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{restore_coordinates}
\alias{restore_coordinates}
\title{restore_coordinates}
\usage{
restore_coordinates(X, mu, sigma)
}
\arguments{
\item{X}{standardized point set as N x D matrix}

\item{mu}{1 x D vector of means}

\item{sigma}{standard deviation}
}
\value{
N X D matrix of unstandardized coordinates
}
\description{
Restore coordinates from mean 0 and standard deviation 1
to their original distribution
}
