% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotx}
\alias{rotx}
\title{rotx}
\usage{
rotx(theta)
}
\arguments{
\item{theta}{angle in radians}
}
\value{
a 3x3 rotation matrix
}
\description{
Create a rotation matrix representing a rotation of theta radians about the x-axis
}
