% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{scale_alpha_shape}
\alias{scale_alpha_shape}
\title{scale_alpha_shape}
\usage{
scale_alpha_shape(as, s)
}
\arguments{
\item{as}{an alpha-shape object of class ashape3d}

\item{s}{scaling factor}
}
\value{
an object of class ashape3d
}
\description{
Uniformly scale an alpha-shape.
Note that this computes the alpha-shape of the scaled point set 
associated with the input alpha-shape.
}
