% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{standardize_coordinates}
\alias{standardize_coordinates}
\title{standardize_coordinates}
\usage{
standardize_coordinates(X)
}
\arguments{
\item{X}{point set as N x D matrix}
}
\value{
a list of 
     X: standardized matrix, 
     mu: vector of means, 
     sigma: standard deviation
}
\description{
Transform coordinates to have mean 0 and standard deviation 1
}
